## What this PR does

This PR improves the performance and maintainability of the admin survey question group interface by:

- Eliminating N+1 queries across question group rendering.

- Preloading and grouping related data (Survey, SurveyAssignment, Rapidfire::Question, and PaperTrail::Version) for efficient access in views.

- Refactoring helper methods to accept preloaded data and reduce query load during rendering.

## Changes

- **View (`index.html.haml`)**
   - Preloaded question group-related associations: survey mappings, surveys with assignments, related questions, and version history.
   - Passed preloaded data to partials to reduce query repetition.

- **Partial (`_question_group.html.haml`)**
   - Switched to using preloaded data for questions, status, and author display.
   - Ensured fallback behavior remains consistent.

- **Helper (`SurveysAnalyticsHelper`)**
  - Added `question_group_survey_author` to retrieve author from a version object.
  - Updated `question_group_status` to accept grouped objects instead of fetching them dynamically.


## Where is the N+1 this fixes, and what do the N+1 queries look like

**a) Where is the N+1 this fixes**


https://github.com/user-attachments/assets/af514b95-f7b3-4719-a706-018193e841eb



**b) what do the N+1 queries look like**


https://github.com/user-attachments/assets/cb49f42c-2ae4-462e-8d36-a45019623d45



## Rails Log Performance Comparison (Average)

| Metric                | Before (N+1)      | After (Optimized) | Improvement         |
| --------------------- | ----------------- | ----------------- | ------------------- |
| **Avg Response Time** | ~218 ms           | ~63 ms            | ~71% faster         |
| **Avg DB Time**       | ~82 ms            | ~18 ms            | ~78% reduction      |
| **Avg Allocations**   | ~68,000+ objects  | ~27,000 objects   | ~60% fewer objects  |

![Screenshot from 2025-06-13 15-05-18](https://github.com/user-attachments/assets/5cf93860-18a9-4afc-99f2-75f2278df537)


## Screenshots
Before:


https://github.com/user-attachments/assets/c17f2ac2-0da6-46b9-8938-bf0396be6f51


![Before Screenshot](https://github.com/user-attachments/assets/ff987a36-e0e9-4199-ab10-38568df116a8)


![Before Rails Log](https://github.com/user-attachments/assets/248abb7c-5dac-4abf-a6ee-70ea7c74dc88)



After:


![After Rails Log](https://github.com/user-attachments/assets/991fd9ba-86c6-42f7-8659-766382eb37af)

![After Screenshot](https://github.com/user-attachments/assets/a3d0ca9e-a971-4df5-a46f-4503737a335b)


https://github.com/user-attachments/assets/0ba4d319-71aa-48de-856a-7eade0c5d3a5