@ragesoss 2025-06-13
This looks good to me. I will merge once the tests pass.